Allow dot-traversal access to nested dictionaries.
A simple memoized recursion has been added to `get()` if the key
contains at least one dot.

The caller can choose to opt-out of this behaviour by specifying the
`dotted_lookup` argument:

```python
settings('AUTH.USERNAME', dotted_lookup=False)
```

While the performance impact of this has not been quantified, the net
impact in any real-world application should be minimal due to typical
nesting levels, and the fact that we overwrite the memoized portion of
the dotted-key lookup on each iteration.

- Avoids regressions [✓]
- Can be opted-out on a per-call basis [✓]
- Minimal performance impact [✓]
- Documented [✓]
- Tested [✓]
- Examples added [✓]

Closes rochacbruno/dynaconf#84