Implements dotted-path validator name declarations.
One is now able to write:

```python
settings.validators.register(
    Validator('REDIS',  must_exist=True,  is_type_of=dict),
    Validator('REDIS.HOST', must_exist=True, is_type_of=str),
    Validator('REDIS.PORT', must_exist=True, is_type_of=int),
)
```

Which will validate the dotted attributes as nested structures. For
example, in yaml:

```yaml
DEFAULT:
    REDIS:
        HOST: localhost
        PORT: 1234
```

This necessitated a slight but non-negligible change in the
implementation of `Settings.exists()`, which previously did a shallow
check of loaded data. It has now been updated to perform a
`Settings.get()` of the key in question, and compares that to a newly
defined sentinel value to ensure `None` values do not cause a false
negative result.

New tests and assertions have been added to cover the new functionality.
Docs have been updated to show an example of the nested validator name
definition in action.

Closes rochacbruno/dynaconf#85.