feat: Add mkdocstrings support with autogen API reference
Followed the recipe on mkdocstrings docs to setup the autogeneration
of API reference docs. Had to make some adjustments because this project
does not have a src/ folder, just the folder with the project name.

Tested results with mkdocs serve and everything builds. It takes ~80
seconds to generate the whole API reference. Some exclusions may be
worth introducing.