docs: Include snippet from code base for PrefixFilter
This change pulls the source code directly from the code base. I think
this change is just a demo and will probably be reversed in favour of
@pedro-psb suggestion in #1059 where we link to a new section in the
reference. Waiting for that change to be merged first.

This method is also quite brittle if the line numbers change. Snippet
sections can be specified, but better to use mkdocstrings.
https://facelessuser.github.io/pymdown-extensions/extensions/snippets/#snippet-sections