docs: Add docstrings to PrefixFilter and include in ref
This change creates a new section in the reference: dynaconf.strategies.
In configuration.md under filtering strategies, we now link to the
generated docs for PrefixFilter. In the docstrings for PrefixFilter we
create a reference to the "Prefix filtering" section in Advanced Usage.

Added a few new options as part of this.