chore(ci): Replace lint and formatting tools with ruff (#1074)
* ci: Add initial ruff config to match existing

* ci: Replace pre-commit hooks with ruff

Replaces black, flake8 and pyupgrade. The flake8 plugins are covered in
the ruff.toml.

* ci: Make ruff generate Py3.7 compatible code

* ci: Replace reorder-python-imports with ruff

* ci: Add case insensitive sorting to ruff isort

* ci: Make sure the import of annotations is preserved

This was being offered by the pre-commit hook which has just been
removed: reorder-imports...

* ci: Add isort rules to lint and put order-by-type=false

This is closest to how things were before.

* ci: Run ruff over dynaconf/

* ci: Run ruff over tests/

* ci: Run ruff format on dynaconf/ again

* ci: Make test_functional ruff.toml extend main one

Only ignores print warnings now.

* ci: Run ruff over tests_functional/

* ci: Remove debug hook because its covered by ruff T100

* ci: Remove encoding declaration hook

Covered by ruff UP rules:
https://docs.astral.sh/ruff/rules/utf8-encoding-declaration/

* ci: Remove more hooks covered by ruff

* Fix comment

* ci: Run pre-commit over all files

* ci: Remove flake8 packages from requirements_dev

* ci: Add support for pep8-naming in ruff

* ci: Ruff fixing import groups

Not sure about this one since it treats dynaconf and pytest as part of
the same group, perhaps because it's not part of the package?

* ci: Remove .pep8speaks - no need for it with ruff

* ci: Remove pep8 section from Makefile

Doesn't need to be replaced with ruff since it will run as part of the
pre-commit section.

* Add .python-version from pyenv to .gitingore

* ci: Add dynaconf as known first party package

* ci: Final fix from ruff

* chore: Remove flake8 test-requirements from setup.py

* chore: Remove flake8 config in setup.cfg

* ci: Bump ruff lint to python v3.8

* Add missing import of annotations

* Replace lower level ruff.toml with per-file-ignores

Everything can now be controlled from one file. And it resolves
potential issues by having the whole "lint" section overwritten. For
example, by changing this it picked up that the imports weren't being
sorted in one file, since this section was overridden in the lower
config file.

* ci: Move E402 ignore to tests_functional/ only and apply fixes

* Move F401 to per-file-ignores and do fixes

* Add noqa on unused imports in TYPE_CHECKING sections

* Remove check for E401

It's already enforced by urff under the [lint.isort] settings

* Add per-file-ignores to __init__.py for imports

Meant that line-by-line blanket # noqa pragmas could be removed.

* Removed F841 - can be added to per-file-ignores as exception

---------

Co-authored-by: Pedro Brochado <pedropsb95@gmail.com>