fix(hooking):Don't instantiate a Settings object for every hook call.
SCENARIO:

With a `function` as `_registered_hooks` every time a key is accessed on `settings`
the `function` is invoked passing a `settings` object as first argument.

BEFORE:

The `settings` passed to the hook was instantiated for every call.

AFTER:

The `settings` is now a `TempSettingsHolder` that has no effect on passing
and is instantiated only if accessed.

Saved execution time from 0m49 to 0m3 on a Django openapi spec view.