feat: add Settings-level caching mechanism
This consolidates a caching mechanism that using a global cache dict and
keys that leverages unique ids of Setting instances. This is only
for the top level Setting object.

The cache addopts a strategy of brute force invalidation. A lot of
methods that may get the cache in an invalid state just clears the
entire cache for all instances. Maybe we 'optimize' this, but caching
invalidation is tricky. Better play safe.